# swagger.yaml
openapi: 3.0.0
info:
  title: Ani & Ayu API
  description: E-commerce API for children's ethnic wear
  version: 1.0.0
  contact:
    name: API Support
    email: support@aniayu.com
servers:
  - url: http://localhost:3001
    description: Development server
  - url: https://api.aniayu.com
    description: Production server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Enter JWT token in format - Bearer <token>
  schemas:
...
    CartItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantity:
          type: integer
        size:
          type: string
        color:
          type: string
        product:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            price:
              type: number
            image_url:
              type: string
            slug:
              type: string
        total:
          type: number
...
paths:
...
  /api/cart:
    get:
      summary: Get current cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: x-guest-id
          in: header
          schema:
            type: string
          description: Guest session ID
      responses:
        '200':
          description: Cart retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/CartItem'
                  summary:
                    type: object
                    properties:
                      totalItems:
                        type: integer
                      subtotal:
                        type: number
  /api/cart/add:
    post:
      summary: Add item to cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: x-guest-id
          in: header
          schema:
            type: string
          description: Guest session ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
              properties:
                productId:
                  type: string
                  format: uuid
                quantity:
                  type: number
                  minimum: 1
                size:
                  type: string
                color:
                  type: string
      responses:
        '200':
          description: Item added to cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  item:
                    type: object
  /api/cart/item/{id}:
    delete:
      summary: Remove item from cart
      tags:
        - Cart
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: x-guest-id
          in: header
          schema:
            type: string
          description: Guest session ID
      responses:
        '200':
          description: Item removed from cart
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string

  # Address Endpoints
...